Added attributes to WorkingMemory. Added AttributeKind and its constants, as well as Bag and its subtypes, to help with type-safety of attributes of WorkingMemory. Added BELIEF_SYSTEM_GRAPH of type ConceptGraph as an attribute to WorkingMemory. Removed beliefSystemGraph argument from method DeriverNodeFactory#createDeriverNode. Replaced properties willingToDeviateContributors and unwillingToDeviateContributors on BelieverDeviationDeriverNode by attributes on WorkingMemory.

+++ EpistemicsOfTheVirtual/MentalWorld/EngineDeriver/src/main/java/selemca/epistemics/mentalworld/engine/deriver/matcher/CategoryMatchImpl.java
@@ -12,9 +12,6 @@ import selemca.epistemics.mentalworld.engine.realitycheck.RealityCheck;
 
 import java.util.*;
 
-/**
- * Created by henrizwols on 06-03-15.
- */
 public class CategoryMatchImpl implements CategoryMatch {
     private final Concept concept;
     private final RealityCheck realityCheck;
@@ -45,14 +42,14 @@ public class CategoryMatchImpl implements CategoryMatch {
     public double getMatchScore() {
         double inverseScore = 1.0;
         for (Double truthValue : contributions.values()) {
-            inverseScore *= (1.0 - truthValue.doubleValue());
+            inverseScore *= (1.0 - truthValue);
         }
         return 1.0 - inverseScore;
     }
 
     public double getContributorScore(Concept contributor) {
         Double score = contributions.get(contributor);
-        return score == null ? 0 : score.doubleValue();
+        return score == null ? 0 : score;
     }
 
     public boolean withinReality(Concept contributor) {
